+++ test/src/Game/Werewolf/Test/Player.hs
@@ -1,0 +1,23 @@
+{-|
+Module      : Game.Werewolf.Test.Player
+Copyright   : (c) Henry J. Wylde, 2015
+License     : BSD3
+Maintainer  : public@hjwylde.com
+-}
+
+{-# OPTIONS_HADDOCK hide, prune #-}
+
+module Game.Werewolf.Test.Player (
+    prop_newPlayerIsAlive,
+) where
+
+import Control.Lens
+
+import Data.Text
+
+import Game.Werewolf.Player
+import Game.Werewolf.Role
+import Game.Werewolf.Test.Arbitrary ()
+
+prop_newPlayerIsAlive :: Text -> Role -> Bool
+prop_newPlayerIsAlive name role = newPlayer name role ^. state == Alive