+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## Changelog
 
 #### Upcoming
+
+*Major*
+
+* Initial implementation with villagers and werewolves.